Add a new feature, bug bonuses.
Every time your length is divisible by 5 a bug, randomly chosen
from a tuple in constants.py, appears on the screen.
In the "statusbar" appears a counter that counts down from a random
number between 20 and 30.
If the counter reaches 0, the bug disappears, but if you eat the bug
before it times out, then the remaining time is added to your score.