Change menu design to use arrow key menus
Instead of hotkey menus, use menus that you navigate with the arrow
keys. The enter key is unreliable, so if it doesn't work, use space.

Some minor changes have been made that are not part of the major update,
things that I noticed while looking at the program, and playing the
game.
They are:
  - The bug timeout is on average a little longer
  - The constants.py docstring has been properly updated
  - Probably some other minor things I've forgotten